+++ internal/client/entity/job_definition/output_options/bigquery.go
