+++ internal/client/job_definition.go
@@ -3,11 +3,11 @@ package client
 import (
 	"fmt"
 	"net/http"
-	jobDefinitionEntities "terraform-provider-trocco/internal/client/entities/job_definitions"
-	filterEntities "terraform-provider-trocco/internal/client/entities/job_definitions/filter"
-	inputOptionEntitites "terraform-provider-trocco/internal/client/entities/job_definitions/input_options"
-	outputOptionEntitites "terraform-provider-trocco/internal/client/entities/job_definitions/output_options"
 	"terraform-provider-trocco/internal/client/entity"
+	jobDefinitionEntities "terraform-provider-trocco/internal/client/entity/job_definition"
+	filterEntities "terraform-provider-trocco/internal/client/entity/job_definition/filter"
+	inputOptionEntitites "terraform-provider-trocco/internal/client/entity/job_definition/input_options"
+	outputOptionEntitites "terraform-provider-trocco/internal/client/entity/job_definition/output_options"
 	"terraform-provider-trocco/internal/client/parameter"
 	"terraform-provider-trocco/internal/client/parameter/job_definitions"
 	filterParameters "terraform-provider-trocco/internal/client/parameter/job_definitions/filter"
