+++ internal/client/job_definition_test.go
@@ -4,9 +4,9 @@ import (
 	"encoding/json"
 	"net/http"
 	"net/http/httptest"
-	"terraform-provider-trocco/internal/client/entities/job_definitions"
-	"terraform-provider-trocco/internal/client/entities/job_definitions/filter"
 	"terraform-provider-trocco/internal/client/entity"
+	job_definitions "terraform-provider-trocco/internal/client/entity/job_definition"
+	"terraform-provider-trocco/internal/client/entity/job_definition/filter"
 	"terraform-provider-trocco/internal/client/parameter"
 	filter2 "terraform-provider-trocco/internal/client/parameter/job_definitions/filter"
 	"testing"
