+++ internal/provider/model/job_definition/input_option/parser/csv_parser.go
@@ -1,7 +1,7 @@
 package parser
 
 import (
-	"terraform-provider-trocco/internal/client/entities/job_definitions"
+	job_definitions "terraform-provider-trocco/internal/client/entity/job_definition"
 	params "terraform-provider-trocco/internal/client/parameter/job_definitions"
 	"terraform-provider-trocco/internal/provider/model"
